--- naivepyrunner/handler.py
@@ -1,6 +1,0 @@
-class Handler():
-    def execute():
-        return -1
-
-    def stop():
-        pass
